SSCS-3900 Do not send notifications out of hours
We are using the scheduler service to schedule notifications that are
received our of hours. This means we will be scheduling COH and CCD
notifications where as before it was just CCD. The COH notifications
are made up of a case_id and coh_reference. Therefore we need to store
different things in the job details payload.

Added the ability to select a different executor and deserializer based
on the payload. This is done in the JobMapper class which will select
the executor and deserializer based on a predicate which tests the
payload.